Update colonel blotto env for early stopping.
Unless rounds_needed_to_win was not reached, the game would continue. Eg: if there were 4 wins for A and 2 ties, but there are 9 rounds, B cannot possibly win. However based on current logic, A won't be declared winner unless it wins the 5th round. So this results in an earlier stop.